+++ app/src/main/java/com/bookshelfhub/bookshelfhub/helpers/cloudstorage/ICloudStorage.kt
@@ -11,7 +11,7 @@ interface ICloudStorage {
         subfolder:String,
         fileName:String,
         remoteFileExt:String,
-        onProgress:(Long)->Unit,
+        onProgress:(Int)->Unit,
         onComplete:()->Unit,
         onError:(Exception)->Unit
     ): FileDownloadTask.TaskSnapshot?
